+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Version 0.0.7 (wip)
-- ...
+- load max icon in fleet overview page
 
 # Version 0.0.6
 - Refactored scan page logic to use jsonPageData (+ small fixes on exports)